requirements.txt
pandas
xgboost
scikit-learn
imblearn
shap
numpy
tabpfn

matplotlib
seaborn

faiss-cpu
sentence-transformers
openai

requests
streamlit
PyPDF2
python-dotenv

ipykernel
jupyter
notebook
